refactor: cov extends test & mv helper fn to command
