+++ lib/cmd/cov.js
@@ -1,12 +1,15 @@
 /* istanbul ignore next */
 'use strict';
 
+const debug = require('debug')('egg-bin:cov');
 const path = require('path');
 const mkdirp = require('mz-modules/mkdirp');
 const rimraf = require('mz-modules/rimraf');
-const Command = require('../command');
+
+const Command = require('./test');
 const EXCLUDES = Symbol('cov#excludes');
 
+/* istanbul ignore next */
 class CovCommand extends Command {
   constructor(argv) {
     super(argv);
@@ -60,6 +63,7 @@ class CovCommand extends Command {
 
     // save coverage-xxxx.json to $PWD/coverage
     const covArgs = this.getCovArgs(context);
+    debug('covArgs: %j', covArgs);
     yield this.helper.forkNode(covFile, covArgs, opt);
     yield rimraf(tmpDir);
 
@@ -95,7 +99,8 @@ class CovCommand extends Command {
       covArgs.push(exclude);
     }
     const mochaFile = require.resolve('mocha/bin/_mocha');
-    const testArgs = this.helper.formatTestArgs(argv);
+    const testArgs = this.formatTestArgs(argv);
+    debug('testArgs: %j', testArgs);
     return covArgs.concat(mochaFile, '--', ...testArgs);
   }
