+++ lib/cmd/dev.js
@@ -2,12 +2,18 @@
 
 const debug = require('debug')('egg-bin:dev');
 const Command = require('../command');
+const path = require('path');
+const utils = require('egg-utils');
+const detect = require('detect-port');
 
 class DevCommand extends Command {
-
   constructor(rawArgv) {
     super(rawArgv);
     this.usage = 'Usage: egg-bin dev [dir] [options]';
+
+    this.defaultPort = 7001;
+    this.serverBin = path.join(__dirname, '../start-cluster');
+
     this.options = {
       baseDir: {
         description: 'directory of application, default to `process.cwd()`',
@@ -37,16 +43,59 @@ class DevCommand extends Command {
   * run(context) {
     const { cwd, argv, execArgv } = context;
     const framework = this.getFrameworkOrEggPath(context);
-    const devArgs = yield this.helper.formatArgs(cwd, argv, { framework });
+    const devArgs = yield this.formatArgs(cwd, argv, { framework });
     const options = {
       execArgv,
       env: Object.assign({ NODE_ENV: 'development' }, process.env),
     };
-    debug('%s %j %j, %j', this.helper.serverBin, devArgs, options.execArgv, options.env.NODE_ENV);
-    yield this.helper.forkNode(this.helper.serverBin, devArgs, options);
+    debug('%s %j %j, %j', this.serverBin, devArgs, options.execArgv, options.env.NODE_ENV);
+    yield this.helper.forkNode(this.serverBin, devArgs, options);
   }
 
+  // keep, will be override
   getFrameworkOrEggPath() {}
+
+  /**
+   * format egg startCluster args then change it to json string style
+   * @method helper#formatArgs
+   * @param {String} cwd - base directory
+   * @param {Object} argv - yargs style
+   * @param {Object} [options] - other options
+   * @param {String} [options.framework] - egg framework name or path
+   * @return {Array} pass to start-cluster, [ '{"port":7001,"framework":"egg"}' ]
+   */
+  * formatArgs(cwd, argv, options = {}) {
+    /* istanbul ignore next */
+    argv.baseDir = argv._[0] || argv.baseDir || cwd;
+    /* istanbul ignore next */
+    if (!path.isAbsolute(argv.baseDir)) argv.baseDir = path.join(cwd, argv.baseDir);
+
+    argv.workers = argv.cluster || 1;
+    argv.port = argv.port || argv.p;
+    argv.framework = utils.getFrameworkPath({
+      framework: argv.framework || options.framework,
+      baseDir: argv.baseDir,
+    });
+
+    // remove unused properties
+    argv.cluster = undefined;
+    argv.c = undefined;
+    argv.p = undefined;
+    argv._ = undefined;
+    argv.$0 = undefined;
+
+    // auto detect available port
+    if (!argv.port) {
+      debug('detect available port');
+      const port = yield detect(this.defaultPort);
+      if (port !== this.defaultPort) {
+        argv.port = port;
+        console.warn(`[egg-bin] server port ${this.defaultPort} is in use, now using port ${port}\n`);
+      }
+      debug(`use available port ${port}`);
+    }
+    return [ JSON.stringify(argv) ];
+  }
 }
 
 module.exports = DevCommand;
