+++ lib/helper.js
@@ -1,110 +1,7 @@
 'use strict';
 
-const fs = require('fs');
-const path = require('path');
-const globby = require('globby');
-const detect = require('detect-port');
-const debug = require('debug')('egg-bin');
-const utils = require('egg-utils');
 const unparse = require('dargs');
 const changeCase = require('change-case');
-
-exports.defaultPort = 7001;
-exports.serverBin = path.join(__dirname, 'start-cluster');
-
-/**
- * format test args then change it to array style
- * @method helper#formatTestArgs
- * @param {Object} argv - yargs style
- * @return {Array} [ '--require=xxx', 'xx.test.js' ]
- */
-exports.formatTestArgs = argv => {
-  const newArgv = Object.assign({}, argv);
-
-   /* istanbul ignore next */
-  newArgv.timeout = newArgv.timeout || process.env.TEST_TIMEOUT || '30000';
-  newArgv.reporter = newArgv.reporter || process.env.TEST_REPORTER;
-
-  // collect require
-  let requireArr = newArgv.require || newArgv.r || [];
-   /* istanbul ignore next */
-  if (!Array.isArray(requireArr)) requireArr = [ requireArr ];
-
-  requireArr.push(require.resolve('co-mocha'));
-
-  if (requireArr.includes('intelli-espower-loader')) {
-    console.warn('[egg-bin] don\'t need to manually require `intelli-espower-loader` anymore');
-  } else {
-    requireArr.push(require.resolve('intelli-espower-loader'));
-  }
-
-  newArgv.require = requireArr;
-
-  // collect test files
-  let files = newArgv._.slice();
-  if (!files.length) {
-    files = [ process.env.TESTS || 'test/**/*.test.js' ];
-  }
-  // expand glob and skip node_modules and fixtures
-  files = globby.sync(files.concat('!test/**/{fixtures, node_modules}/**/*.test.js'));
-
-  // auto add setup file as the first test file
-  const setupFile = path.join(process.cwd(), 'test/.setup.js');
-  if (fs.existsSync(setupFile)) {
-    files.unshift(setupFile);
-  }
-  newArgv._ = files;
-
-  // remove alias
-  newArgv.$0 = undefined;
-  newArgv.r = undefined;
-  newArgv.t = undefined;
-  newArgv.g = undefined;
-
-  return unparse(newArgv);
-};
-
-/**
- * format egg startCluster args then change it to json string style
- * @method helper#formatArgs
- * @param {String} cwd - base directory
- * @param {Object} argv - yargs style
- * @param {Object} [options] - other options
- * @param {String} [options.framework] - egg framework name or path
- * @return {Array} pass to start-cluster, [ '{"port":7001,"framework":"egg"}' ]
- */
-exports.formatArgs = function* (cwd, argv, options = {}) {
-  /* istanbul ignore next */
-  argv.baseDir = argv._[0] || argv.baseDir || cwd;
-  /* istanbul ignore next */
-  if (!path.isAbsolute(argv.baseDir)) argv.baseDir = path.join(cwd, argv.baseDir);
-
-  argv.workers = argv.cluster || 1;
-  argv.port = argv.port || argv.p;
-  argv.framework = utils.getFrameworkPath({
-    framework: argv.framework || options.framework,
-    baseDir: argv.baseDir,
-  });
-
-  // remove unused properties
-  argv.cluster = undefined;
-  argv.c = undefined;
-  argv.p = undefined;
-  argv._ = undefined;
-  argv.$0 = undefined;
-
-  // auto detect available port
-  if (!argv.port) {
-    debug('detect available port');
-    const port = yield detect(exports.defaultPort);
-    if (port !== exports.defaultPort) {
-      argv.port = port;
-      console.warn(`[egg-bin] server port ${exports.defaultPort} is in use, now using port ${port}\n`);
-    }
-    debug(`use available port ${port}`);
-  }
-  return [ JSON.stringify(argv) ];
-};
 
 /**
  * extract argv then change it to array style
